requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
colorama==0.4.4
Django==4.0
django-braces==1.15.0
django-crispy-forms==1.13.0
isort==5.10.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.12.2
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
wrapt==1.13.3
